criada consulta do Para
